Move all WebCore API calls to separate class
